fix: daily_stats avg_duration 타입 불일치 수정 (Double → Long)
DB avg_duration 컬럼이 int(11)인데 엔티티가 Double로 선언되어 있던 문제 수정.
ms 단위 정수값이므로 Long이 적절하며, AVG() 쿼리 결과는 longValue()로 변환.

+++ backend/src/main/java/com/dashboard/backend/domain/DailyStat.java
@@ -28,7 +28,7 @@ public class DailyStat {
 
     private Long uniqueVisitors;
 
-    private Double avgDuration;
+    private Long avgDuration;
 
     private Double bounceRate;
 
@@ -36,7 +36,7 @@ public class DailyStat {
     private String delYn = "N";
 
     public DailyStat(Project project, LocalDate statDate, Long totalViews,
-                     Long uniqueVisitors, Double avgDuration, Double bounceRate) {
+                     Long uniqueVisitors, Long avgDuration, Double bounceRate) {
         this.project = project;
         this.statDate = statDate;
         this.totalViews = totalViews;
